WIP- Refactor plugins to have more meaningful names, add createDataInfo in s3 storage's stat method
