+++ src/plugins/ImportArtifact/ImportArtifact.js
@@ -2,92 +2,72 @@
 /*eslint-env node, browser*/
 
 define([
-    'deepforge/storage/index',
+    'plugin/PluginConfig',
     'text!./metadata.json',
     'plugin/PluginBase',
-    'deepforge/plugin/Artifacts',
+    'deepforge/storage/index',
+    'deepforge/plugin/Artifacts'
 ], function (
-    Storage,
+    PluginConfig,
     pluginMetadata,
     PluginBase,
-    Artifacts
-) {
+    Storage,
+    Artifacts) {
     'use strict';
 
     pluginMetadata = JSON.parse(pluginMetadata);
-
-    /**
-     * Initializes a new instance of ImportArtifact.
-     * @class
-     * @augments {PluginBase}
-     * @classdesc This class represents the plugin ImportArtifact.
-     * @constructor
-     */
-    var ImportArtifact = function () {
+    const ImportArtifact = function () {
         // Call base class' constructor.
         PluginBase.call(this);
         this.pluginMetadata = pluginMetadata;
     };
 
-    /**
-     * Metadata associated with the plugin. Contains id, name, version, description, icon, configStructue etc.
-     * This is also available at the instance at this.pluginMetadata.
-     * @type {object}
-     */
     ImportArtifact.metadata = pluginMetadata;
 
     // Prototypical inheritance from PluginBase.
     ImportArtifact.prototype = Object.create(PluginBase.prototype);
     Object.assign(ImportArtifact.prototype, Artifacts.prototype);
+
     ImportArtifact.prototype.constructor = ImportArtifact;
 
-    /**
-     * Main function for the plugin to execute. This will perform the execution.
-     * Notes:
-     * - Always log with the provided logger.[error,warning,info,debug].
-     * - Do NOT put any user interaction logic UI, etc. inside this method.
-     * - callback always has to be called even if error happened.
-     *
-     * @param {function(string, plugin.PluginResult)} callback - the result callback
-     */
     ImportArtifact.prototype.main = async function (callback) {
         const config = this.getCurrentConfig();
-        const hash = config.dataHash;
-        const baseName = config.dataTypeId;
         const base = this.getBaseNode();
+        const path = config.dataPath;
+        const baseName = config.dataTypeId;
 
         if (!base) {
             callback(`Could not find data type "${baseName}"`, this.result);
             return;
         }
+
+        const parent = await this.getArtifactsDir();
+        const dataNode = this.core.createNode({base, parent});
+
+        const name = await this.getAssetNameFromPath(path) ||
+            baseName[0].toLowerCase() + baseName.substring(1);
+        let assetInfo;
+
         try {
-            const parent = await this.getArtifactsDir();
-            const dataNode = this.core.createNode({base, parent});
 
-            const name = await this.getAssetNameFromHash(hash) ||
-                baseName[0].toLowerCase() + baseName.substring(1);
-            let assetInfo;
-
-            assetInfo = await this.transfer(hash, config.storage, name);
-
-            this.assignAssetAttributes(dataNode, {type: baseName, name: name, data: assetInfo});
-            await this.save(`Uploaded "${name}" data`);
+            assetInfo = await this.symLink(path, config.storage);
+            this.assignAssetAttributes(dataNode, {data: assetInfo, name: name, type: baseName});
+            await this.save(`Transferred "${name}" data`);
             this.result.setSuccess(true);
             callback(null, this.result);
         } catch (err) {
             callback(err, this.result);
         }
+
     };
 
-    ImportArtifact.prototype.transfer = async function (hash, storage, name) {
-        const filename = `${this.projectId}/artifacts/${name}`;
-        const gmeStorageClient = await Storage.getBackend('gme').getClient(this.logger);
-        const dataInfo = gmeStorageClient.createDataInfo(hash);
-        const content = await gmeStorageClient.getFile(dataInfo);
-
+    ImportArtifact.prototype.symLink = async function(path, storage) {
         const {id, config} = storage;
-        const dstStorage = await Storage.getBackend(id).getClient(this.logger, config);
-        return await dstStorage.putFile(filename, content);
+        if(path.includes(`${this.projectId}/artifacts`)){
+            throw new Error('Cannot import from project root directory');
+        }
+        const srcStorage = await Storage.getBackend(id).getClient(this.logger, config);
+        return await srcStorage.stat(path);
     };
 
     return ImportArtifact;
