--- src/plugins/ImportStoragePaths/metadata.json
@@ -1,32 +1,0 @@
-{
-  "id": "ImportStoragePaths",
-  "name": "ImportStoragePaths",
-  "version": "0.1.0",
-  "description": "",
-  "icon": {
-    "class": "glyphicon glyphicon-cog",
-    "src": ""
-  },
-  "disableServerSideExecution": false,
-  "disableBrowserSideExecution": false,
-  "writeAccessRequired": false,
-  "configStructure": [
-    {
-      "name": "dataPath",
-      "displayName": "Path",
-      "description": "Path to the already existing artifact in the storage backend",
-      "value": "",
-      "valueType": "str",
-      "readOnly": false,
-      "required": true
-    },
-    {
-      "name": "dataTypeId",
-      "displayName": "Data Type",
-      "description": "",
-      "value": "",
-      "valueType": "str",
-      "readOnly": false
-    }
-  ]
-}
